requirements.txt
apache-airflow==2.9.0
pandas==2.2.0
great-expectations==0.18.10
pytest==8.0.0
psycopg2-binary==2.9.9